show local variables, implemented stepping over lines

+++ src/smthelusive/debyter/constants/ModKind.java
@@ -1,16 +1,16 @@
 package smthelusive.debyter.constants;
 
 public class ModKind {
-    public static final int COUNT = 1;
-    public static final int CONDITIONAL = 2;
-    public static final int THREAD_ONLY = 3;
-    public static final int CLASS_ONLY = 4;
-    public static final int CLASS_MATCH = 5;
-    public static final int CLASS_EXCLUDE = 6;
-    public static final int LOCATION_ONLY = 7;
-    public static final int EXCEPTION_ONLY = 8;
-    public static final int FIELD_ONLY = 9;
-    public static final int STEP = 10;
-    public static final int INSTANCE_ONLY = 11;
-    public static final int SOURCE_NAME_MATCH = 12;
+    public static final byte COUNT = 1;
+    public static final byte CONDITIONAL = 2;
+    public static final byte THREAD_ONLY = 3;
+    public static final byte CLASS_ONLY = 4;
+    public static final byte CLASS_MATCH = 5;
+    public static final byte CLASS_EXCLUDE = 6;
+    public static final byte LOCATION_ONLY = 7;
+    public static final byte EXCEPTION_ONLY = 8;
+    public static final byte FIELD_ONLY = 9;
+    public static final byte STEP = 10;
+    public static final byte INSTANCE_ONLY = 11;
+    public static final byte SOURCE_NAME_MATCH = 12;
 }
